scripts/requirements.txt
asyncpg
aiohttp
python-dotenv
apscheduler
motor
requests
markdown
